4251: log cleanup; fixing rename of test file
