+++ web-client/integration-tests/journey/associatedUserAdvancedSearchForCase.js
